move return codes tests to own module
