+++ src/borg/testsuite/archiver/__init__.py
@@ -134,24 +134,6 @@ def cmd(request):
     return exec_fn
 
 
-def test_return_codes(cmd, tmpdir):
-    repo = tmpdir.mkdir("repo")
-    input = tmpdir.mkdir("input")
-    output = tmpdir.mkdir("output")
-    input.join("test_file").write("content")
-    rc, out = cmd("--repo=%s" % str(repo), "rcreate", "--encryption=none")
-    assert rc == EXIT_SUCCESS
-    rc, out = cmd("--repo=%s" % repo, "create", "archive", str(input))
-    assert rc == EXIT_SUCCESS
-    with changedir(str(output)):
-        rc, out = cmd("--repo=%s" % repo, "extract", "archive")
-        assert rc == EXIT_SUCCESS
-    rc, out = cmd("--repo=%s" % repo, "extract", "archive", "does/not/match")
-    assert rc == EXIT_WARNING  # pattern did not match
-    rc, out = cmd("--repo=%s" % repo, "create", "archive", str(input))
-    assert rc == EXIT_ERROR  # duplicate archive name
-
-
 """
 test_disk_full is very slow and not recommended to be included in daily testing.
 for this test, an empty, writable 16MB filesystem mounted on DF_MOUNT is required.
